fix/clean up lifecycle spec
staging would trigger multiple times and not be visible because #update!
wasn't being given a callback

changed #update! to never auto-restart; everything told it not to anyway, and
one case assuming it would was wrong, as it was calling the CFoundry #update!,
and not the harness

App#push in the harness will restart if the app is started; this mirrors `cf`