loggregator: runnable against local dev instance
This assumes that if the API endpoint does not use https, the loggregator
server is accessable via port 80 with TCP forwarding and that SSL is not used.
This is the case for bosh-lite, but not e.g. an AWS dev instance.

Given our direction, supporting bosh-lite is probably more useful.